[Fix] Localization 오류 해결
