+++ FlagQuizSwiftUI/FlagQuizSwiftUI.xcodeproj/xcshareddata/xcschemes/FlagQuizSwiftUI.xcscheme
@@ -57,6 +57,7 @@
       buildConfiguration = "Debug"
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
+      language = "ko"
       launchStyle = "0"
       useCustomWorkingDirectory = "NO"
       ignoresPersistentStateOnLaunch = "NO"
